+++ src/models/Settings.php
@@ -2,7 +2,9 @@
 
 namespace vaersaagod\toolmate\models;
 
+use Craft;
 use craft\base\Model;
+use craft\helpers\ConfigHelper;
 
 /**
  * ToolMate Settings Model
@@ -16,20 +18,54 @@ class Settings extends Model
     // Public Properties
     // =========================================================================
 
-    
-    public $publicRoot = null;
+    /** @var string */
+    public $publicRoot = '@webroot';
+
+    /** @var bool */
     public $enableMinify = true;
 
-    // Public Methods
-    // =========================================================================
+    /** @var int|string|bool|null */
+    public $embedCacheDuration = null;
+
+    /** @var int|string|bool|null */
+    public $embedCacheDurationOnError = null;
 
     /**
-     * @return array
+     * @throws \yii\base\InvalidConfigException
      */
-    public function rules(): array
+    public function init()
     {
-        return [
-
-        ];
+        parent::init();
+        $this->setAttributes($this->getAttributes(), false);
     }
+
+    /**
+     * @param array $values
+     * @param bool $safeOnly
+     * @throws \yii\base\InvalidConfigException
+     */
+    public function setAttributes($values, $safeOnly = true)
+    {
+
+        parent::setAttributes($values, $safeOnly);
+
+        $this->publicRoot = Craft::parseEnv($this->publicRoot) ?: ($_SERVER['DOCUMENT_ROOT'] ?? '');
+
+        if ($this->embedCacheDuration !== false) {
+            if ($this->embedCacheDuration !== null) {
+                $this->embedCacheDuration = ConfigHelper::durationInSeconds($this->embedCacheDuration);
+            } else {
+                $this->embedCacheDuration = Craft::$app->getConfig()->getGeneral()->cacheDuration;
+            }
+        }
+
+        if ($this->embedCacheDurationOnError !== false) {
+            if ($this->embedCacheDurationOnError !== null) {
+                $this->embedCacheDurationOnError = ConfigHelper::durationInSeconds($this->embedCacheDurationOnError);
+            } else {
+                $this->embedCacheDurationOnError = 300; // 5 minutes
+            }
+        }
+    }
+
 }